services/api_gateway/requirements.txt
# API Gateway dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP client
httpx>=0.24.0

# Service discovery
nacos-sdk-python>=0.1.14

# Configuration
python-dotenv>=1.0.0
